feat: Separate interview messenger body
