+++ src/components/interview/AlertNewMessage.component.tsx
@@ -11,13 +11,13 @@ interface Props {
 function AlertNewMessageComponent({ contents }: Props) {
   return (
     <div className={styles.container}>
-      <div className={styles.box}>
+      <button type="button" className={styles.box}>
         <p className={styles.title}>새로운 메시지</p>
         <div className={styles.contents}>
           {contents}
           <ArrowDownOutlined />
         </div>
-      </div>
+      </button>
     </div>
   );
 }
