+++ src/styles/components/AlertNewMessage.module.scss
@@ -14,11 +14,15 @@
 }
 
 .box {
+  @extend .buttonWithNoStyle;
+
   border: 1px solid $color-border-gray;
   border-radius: 5px;
   padding: 14px;
   cursor: pointer;
   background-color: white;
+
+  width: 100%;
 
   box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
   .title {
